# web/openapi.json
{
  "components": {
    "schemas": {
...
      "RuleSummaryResponse": {
        "properties": {
          "destination": {
            "$ref": "#/components/schemas/NamedCalendarEndpointResponse"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_sync": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunOutcomeResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviewSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "privacy_policy": {
            "enum": [
              "busy_only",
              "copy_details"
            ],
            "title": "Privacy Policy",
            "type": "string"
          },
          "reprojection_required": {
            "title": "Reprojection Required",
            "type": "boolean"
          },
          "running": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleWorkResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "$ref": "#/components/schemas/NamedCalendarEndpointResponse"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "sync_all_day_events": {
            "title": "Sync All Day Events",
            "type": "boolean"
          },
          "tentative_events": {
            "enum": [
              "sync",
              "mark",
              "skip"
            ],
            "title": "Tentative Events",
            "type": "string"
          },
          "unanswered_invitations": {
            "enum": [
              "wait",
              "as_tentative"
            ],
            "title": "Unanswered Invitations",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "destination",
          "privacy_policy",
          "sync_all_day_events",
          "tentative_events",
          "unanswered_invitations",
          "state",
          "reprojection_required",
          "last_sync",
          "latest_preview",
          "running"
        ],
        "title": "RuleSummaryResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Calendar Ghost",
    "version": "0.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rules": {
      "get": {
        "operationId": "list_rules_api_v1_rules_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleSummaryResponse"
                  },
                  "title": "Response List Rules Api V1 Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rules"
      },
      "post": {
        "operationId": "create_rule_api_v1_rules_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule"
      }
    },
    "/api/v1/rules/{rule_id}": {
      "delete": {
        "operationId": "remove_rule_api_v1_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "projections",
            "required": true,
            "schema": {
              "enum": [
                "delete",
                "detach"
              ],
              "title": "Projections",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Rule"
      },
      "get": {
        "operationId": "rule_details_api_v1_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rule Details"
      },
      "patch": {
        "operationId": "change_rule_policy_api_v1_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRulePolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Rule Policy"
      }
    },
    "/api/v1/rules/{rule_id}/enable": {
      "post": {
        "operationId": "enable_rule_api_v1_rules__rule_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Rule"
      }
    },
    "/api/v1/rules/{rule_id}/pause": {
      "post": {
        "operationId": "pause_rule_api_v1_rules__rule_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Rule"
      }
    },
    "/api/v1/rules/{rule_id}/preview": {
      "post": {
        "operationId": "preview_rule_api_v1_rules__rule_id__preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Rule"
      }
    },
    "/api/v1/rules/{rule_id}/reconcile": {
      "post": {
        "operationId": "reconcile_now_api_v1_rules__rule_id__reconcile_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile Now"
      }
    },
    "/api/v1/rules/{rule_id}/replace": {
      "post": {
        "operationId": "replace_rule_calendars_api_v1_rules__rule_id__replace_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleReplacementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Rule Calendars"
      }
    },
    "/api/v1/rules/{rule_id}/sync": {
      "post": {
        "operationId": "sync_now_api_v1_rules__rule_id__sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Now"
      }
    },
...
  }
}